+++ Ya/Algebra/Definition.hs
@@ -8,17 +8,16 @@ import Ya.Algebra.Abstract
 class Mapping v from into f g where
 	mapping :: v from s t -> into (f s) (g t)
 
-transform :: forall v from into f g s t .
+map :: forall v from into f g s t .
 	Mapping v from into f g =>
 	Castable Dual Arrow (v from s t) =>
 	Supertype (v from s t) -> into (f s) (g t)
-transform from = mapping @v @from @into @f @g @s @t (wrap @Arrow from)
+map from = mapping @v @from @into @f @g @s @t (wrap @Arrow from)
 
 component :: forall v from into f g t .
-	Category from =>
+	Component v from into f g =>
 	(Supertype (v from t t) ~ from t t) =>
 	Castable Dual Arrow (v from t t) =>
-	Mapping v from into f g =>
 	into (f t) (g t)
 component = mapping @v @from @into @f @g @_ @t (wrap @Arrow identity)
 
@@ -28,7 +27,7 @@ class
 	, forall i . Mapping Dual from Arrow (U_II_I from i) (U_II_I from i)
 	) => Precategory from where
 	compose :: from s t -> from i s -> from i t
-	compose post pre = let U_I_II y = transform @Flat post (U_I_II pre) in y
+	compose post pre = let U_I_II y = map @Flat post (U_I_II pre) in y
 
 deriving instance
 	( forall i . Mapping Flat from Arrow (U_I_II from i) (U_I_II from i)
@@ -40,38 +39,47 @@ deriving instance
 class Precategory from => Category from
 	where identity :: from s s
 
+{- [LAW] Identity preserving: mapping identity ≡ identity -}
+{- [LAW] Composition preserving: mapping (f . g) ≡ mapping f . mapping g -}
 class (m from, m into, Mapping v from into f f) => Functor v m from into f
 deriving instance (m from, m into, Mapping v from into f f) => Functor v m from into f
 
-{- [LAW] Composition preserving: mapping (f . g) ≡ mapping f . mapping g -}
+functor :: forall v from into f s t .
+	Functor v Category from into f =>
+	Castable Dual Arrow (v from s t) =>
+	Supertype (v from s t) -> into (f s) (f t)
+functor = map @v @from @into @f @f @s @t
 
+{- [LAW] Composition preserving: mapping (f . g) ≡ mapping f . mapping g -}
 type Semi v x = x v Precategory
 
 semifunctor :: forall v from into f s t .
 	Semi v Functor from into f =>
 	Castable Dual Arrow (v from s t) =>
 	Supertype (v from s t) -> into (f s) (f t)
-semifunctor = transform @v @from @into @f @f @s @t
-
-{- [LAW] Identity preserving: mapping identity ≡ identity -}
-{- [LAW] Composition preserving: mapping (f . g) ≡ mapping f . mapping g -}
-functor :: forall v from into f s t .
-	Functor v Category from into f =>
-	Castable Dual Arrow (v from s t) =>
-	Supertype (v from s t) -> into (f s) (f t)
-functor = transform @v @from @into @f @f @s @t
+semifunctor = map @v @from @into @f @f @s @t
 
 type Endo v x c into = x v c into into
 
-class (m from, m into, Mapping v from into f g, Mapping v from into f f, Mapping v from into g g)
-	=> Compositional v m from into f g
-
-deriving instance (m from, m into, Mapping v from into f g, Mapping v from into f f, Mapping v from into g g)
-	=> Compositional v m from into f g
-
 {- LAW: mapping @g @g morphism . component @f @g = component @f @g . mapping morphism @f @f -}
 {- LAW: mapping @g @g morphism . component @f @g = component @f @g . mapping morphism @f @f -}
-type Natural f = f Flat
+class
+	( m from, m into
+	, Mapping v from into f g
+	, Mapping v from into f f
+	, Mapping v from into g g
+	) => Transformation v m from into f g
+
+deriving instance
+	( m from, m into
+	, Mapping v from into f g
+	, Mapping v from into f f
+	, Mapping v from into g g
+	) => Transformation v m from into f g
+
+type Natural = Flat
+
+type Component v = Transformation v Category
 
 type Covariant f = f Flat
 
@@ -99,13 +107,13 @@ type family Representation t where
 	Representation (U_I_I (/\)) = Unit \/ Unit
 
 class
-	 ( Compositional v x from into t (v hom (Representation t))
-	 , Compositional v x from into (v hom (Representation t)) t
+	 ( Transformation v x from into t (v hom (Representation t))
+	 , Transformation v x from into (v hom (Representation t)) t
 	 ) => Representable hom v x from into t
 
 deriving instance
-	( Compositional v x from into t (v hom (Representation t))
-	, Compositional v x from into (v hom (Representation t)) t
+	( Transformation v x from into t (v hom (Representation t))
+	, Transformation v x from into (v hom (Representation t)) t
 	) => Representable hom v x from into t
 
 type family Neutral p where